Fix moved pass credentials to credentials.py
