+++ Babylon/groups/azure/groups/ad/groups/app/commands/get.py
@@ -7,7 +7,7 @@ from rich.pretty import pretty_repr
 from ........utils.request import oauth_request
 from ........utils.response import CommandResponse
 from ........utils.decorators import output_to_file
-from ........utils.decorators import pass_azure_token
+from ........utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
