+++ Babylon/groups/azure/groups/ad/groups/app/groups/password/commands/delete.py
@@ -7,7 +7,7 @@ from click import argument
 from ..........utils.interactive import confirm_deletion
 from ..........utils.request import oauth_request
 from ..........utils.response import CommandResponse
-from ..........utils.decorators import pass_azure_token
+from ..........utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
