+++ Babylon/groups/azure/groups/ad/groups/group/groups/member/commands/add.py
@@ -6,7 +6,7 @@ from click import argument
 from ..........utils.request import oauth_request
 from ..........utils.response import CommandResponse
 from ..........utils.typing import QueryType
-from ..........utils.decorators import pass_azure_token
+from ..........utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
