+++ Babylon/groups/azure/groups/ad/groups/group/groups/member/commands/remove.py
@@ -7,7 +7,7 @@ from click import option
 from ..........utils.request import oauth_request
 from ..........utils.response import CommandResponse
 from ..........utils.interactive import confirm_deletion
-from ..........utils.decorators import pass_azure_token
+from ..........utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
