+++ Babylon/groups/azure/groups/appinsight/commands/delete.py
@@ -5,7 +5,7 @@ from click import command, argument
 from ......utils.response import CommandResponse
 from ......utils.request import oauth_request
 from ......utils.decorators import require_platform_key
-from ......utils.decorators import pass_azure_token
+from ......utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
