+++ Babylon/groups/azure/groups/staticwebapp/commands/delete.py
@@ -8,7 +8,7 @@ from ......utils.request import oauth_request
 from ......utils.decorators import require_platform_key
 from ......utils.response import CommandResponse
 from ......utils.interactive import confirm_deletion
-from ......utils.decorators import pass_azure_token
+from ......utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
