+++ Babylon/groups/azure/groups/staticwebapp/groups/app_settings/commands/get.py
@@ -7,7 +7,7 @@ from rich.pretty import pretty_repr
 from ........utils.request import oauth_request
 from ........utils.decorators import require_platform_key
 from ........utils.response import CommandResponse
-from ........utils.decorators import pass_azure_token
+from ........utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
