+++ Babylon/groups/azure/groups/staticwebapp/groups/app_settings/commands/update.py
@@ -11,7 +11,7 @@ from ........utils.environment import Environment
 from ........utils.request import oauth_request
 from ........utils.decorators import require_platform_key
 from ........utils.response import CommandResponse
-from ........utils.decorators import pass_azure_token
+from ........utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
