+++ Babylon/groups/powerbi/groups/dataset/commands/delete.py
@@ -10,7 +10,7 @@ from ......utils.response import CommandResponse
 from ......utils.request import oauth_request
 from ......utils.typing import QueryType
 from ......utils.interactive import confirm_deletion
-from ......utils.decorators import pass_azure_token
+from ......utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
