+++ Babylon/groups/powerbi/groups/dataset/commands/get.py
@@ -11,7 +11,7 @@ from ......utils.decorators import output_to_file
 from ......utils.response import CommandResponse
 from ......utils.request import oauth_request
 from ......utils.typing import QueryType
-from ......utils.decorators import pass_azure_token
+from ......utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
