+++ Babylon/groups/powerbi/groups/dataset/commands/take_over.py
@@ -9,7 +9,7 @@ from ......utils.decorators import require_deployment_key
 from ......utils.response import CommandResponse
 from ......utils.request import oauth_request
 from ......utils.typing import QueryType
-from ......utils.decorators import pass_azure_token
+from ......utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
