+++ Babylon/groups/powerbi/groups/dataset/commands/update_credentials.py
@@ -8,7 +8,7 @@ from ......utils.response import CommandResponse
 from ......utils.typing import QueryType
 from ......utils.request import oauth_request
 from ......utils.decorators import require_deployment_key
-from ......utils.decorators import pass_azure_token
+from ......utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
