+++ Babylon/groups/powerbi/groups/report/commands/download.py
@@ -10,7 +10,7 @@ from click import Path
 from ......utils.decorators import require_deployment_key
 from ......utils.response import CommandResponse
 from ......utils.request import oauth_request
-from ......utils.decorators import pass_azure_token
+from ......utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
