+++ Babylon/groups/powerbi/groups/report/commands/upload.py
@@ -16,7 +16,7 @@ from ......utils.decorators import timing_decorator
 from ......utils.response import CommandResponse
 from ......utils.request import oauth_request
 from ......utils.typing import QueryType
-from ......utils.decorators import pass_azure_token
+from ......utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
