+++ Babylon/groups/powerbi/groups/workspace/commands/get.py
@@ -8,7 +8,7 @@ from ......utils.logging import table_repr
 from ......utils.response import CommandResponse
 from ......utils.typing import QueryType
 from ......utils.request import oauth_request
-from ......utils.decorators import pass_azure_token
+from ......utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
