+++ Babylon/groups/powerbi/groups/workspace/groups/user/commands/get_all.py
@@ -10,7 +10,7 @@ from ........utils.logging import table_repr
 from ........utils.typing import QueryType
 from ........utils.response import CommandResponse
 from ........utils.request import oauth_request
-from ........utils.decorators import pass_azure_token
+from ........utils.credentials import pass_azure_token
 
 logger = logging.getLogger("Babylon")
